[feature/JENKINS-36211] clean up. move url logic to util class
