+++ blueocean-dashboard/src/main/js/redux/actions.js
@@ -5,65 +5,7 @@ import { State } from '../components/records';
 
 import { getNodesInformation } from '../util/logDisplayHelper';
 
-// helper functions
-
-// helper to clean the path
-function uriString(input) {
-    return encodeURIComponent(input).replace(/%2F/g, '%252F');
-}
-// helper calculate url
-export function calculateLogUrl(config) {
-    if (config.node) {
-        const { nodesBaseUrl, node } = config;
-        return `${nodesBaseUrl}/${node.id}/log`;
-    }
-    return config.url;
-}
-
-export function calculateNodeBaseUrl(config) {
-    const { name, runId, branch, _appURLBase, isMultiBranch } = config;
-    const baseUrl =
-        `${_appURLBase}/rest/organizations/jenkins/` +
-        `pipelines/${uriString(name)}`;
-    if (isMultiBranch) {
-        return `${baseUrl}/branches/${uriString(branch)}/runs/${runId}/nodes/`;
-    }
-    return `${baseUrl}/runs/${runId}/nodes/`;
-}
-
-export function calculateStepsBaseUrl(config) {
-    const { name, runId, branch, _appURLBase, isMultiBranch, node } = config;
-    let baseUrl =
-        `${_appURLBase}/rest/organizations/jenkins/` +
-        `pipelines/${uriString(name)}`;
-    if (isMultiBranch) {
-        baseUrl = `${baseUrl}/branches/${uriString(branch)}`;
-    }
-    // console.log('xxx'), baseUrl;
-    if (node && node !== null) {
-        return `${baseUrl}/runs/${runId}/nodes/${node}/steps`;
-    }
-    return `${baseUrl}/runs/${runId}/steps/`;
-}
-
-export function calculateRunLogURLObject(config) {
-    const { name, runId, branch, _appURLBase, isMultiBranch } = config;
-    const baseUrl = `${_appURLBase}/rest/organizations/jenkins` +
-        `/pipelines/${uriString(name)}`;
-    let url;
-    let fileName;
-    if (isMultiBranch) {
-        url = `${baseUrl}/branches/${uriString(branch)}/runs/${runId}/log/`;
-        fileName = `${branch}-${runId}.txt`;
-    } else {
-        url = `${baseUrl}/runs/${runId}/log/`;
-        fileName = `${runId}.txt`;
-    }
-    return {
-        url,
-        fileName,
-    };
-}
+import {calculateStepsBaseUrl, calculateLogUrl, calculateNodeBaseUrl, calculateRunLogURLObject} from '../util/UrlUtils';
 
 // main actin logic
 export const ACTION_TYPES = keymirror({
@@ -701,7 +643,10 @@ export const actions = {
                 let nodeModel;
                 let node;
                 if (!config.node) {
-                    const focused = information.model.filter((item) => item.isFocused)[0];
+                    const focused = information.model.filter((item) => {
+                        console.log(item)
+                        return item.isFocused;;
+                    })[0];
                     if (focused) {
                         nodeModel = focused;
                     } else {
@@ -720,7 +665,7 @@ export const actions = {
                 return dispatch(actions.fetchSteps(mergedConfig));
             }
 
-            if (!data || !data[nodesBaseUrl]) {
+            if (!data || !data[nodesBaseUrl] || config.refetch) {
                 return exports.fetchJson(
                     nodesBaseUrl,
                     (json) => {
@@ -744,7 +689,7 @@ export const actions = {
         return (dispatch, getState) => {
             const data = getState().adminStore.nodes;
             const nodesBaseUrl = calculateNodeBaseUrl(config);
-            if (!data || !data[nodesBaseUrl]) {
+            if (!data || !data[nodesBaseUrl] || config.refetch) {
                 return actions.fetchNodes(config);
             }
             const node = data[nodesBaseUrl].model.filter((item) => item.id === config.node)[0];
@@ -770,7 +715,7 @@ export const actions = {
         return (dispatch, getState) => {
             const data = getState().adminStore.steps;
             const stepBaseUrl = calculateStepsBaseUrl(config);
-            if (!data || !data[stepBaseUrl] || !data[stepBaseUrl]) {
+            if (!data || !data[stepBaseUrl] || config.refetch) {
                 return exports.fetchJson(
                     stepBaseUrl,
                     (json) => {
